+++ dune/grid/albertagrid/agrid.hh
@@ -164,12 +164,6 @@ namespace Dune
     //! type of communication
     typedef typename Traits::Communication Communication;
 
-    /**
-     * \deprecated Use Communication instead! Will be removed after Dune 2.9.
-     */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
-
   private:
     //! type of LeafIterator
     typedef typename Traits::template Codim<0>::LeafIterator LeafIterator;