+++ dune/grid/albertagrid/gridview.hh
@@ -46,10 +46,6 @@ namespace Dune
     /** \brief type of the communication */
     typedef typename Grid::Traits::Communication Communication;
 
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
-
     template< int cd >
     struct Codim
     {
@@ -97,10 +93,6 @@ namespace Dune
 
     /** \brief type of the communication */
     typedef typename Traits::Communication Communication;
-
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >
@@ -264,10 +256,6 @@ namespace Dune
     /** \brief type of the communication */
     typedef typename Grid::Traits::Communication Communication;
 
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
-
     template< int cd >
     struct Codim
     {
@@ -316,10 +304,6 @@ namespace Dune
 
     /** \brief type of the communication */
     typedef typename Traits::Communication Communication;
-
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >