+++ dune/grid/common/defaultgridview.hh
@@ -42,10 +42,6 @@ namespace Dune
     /** \brief type of the communication */
     typedef typename Grid :: Traits :: Communication Communication;
 
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
-
     template< int cd >
     struct Codim
     {
@@ -96,10 +92,6 @@ namespace Dune
 
     /** \brief type of the communication */
     typedef typename Grid :: Traits :: Communication Communication;
-
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >
@@ -235,10 +227,6 @@ namespace Dune
     /** \brief type of the communication */
     typedef typename Grid :: Traits :: Communication Communication;
 
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
-
     template< int cd >
     struct Codim
     {
@@ -289,10 +277,6 @@ namespace Dune
 
     /** \brief type of the communication */
     typedef typename Grid :: Traits :: Communication Communication;
-
-    /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-    [[deprecated("Use Communication instead!")]]
-    typedef Communication CollectiveCommunication;
 
     /** \brief Codim Structure */
     template< int cd >