+++ dune/grid/uggrid/uggridviews.hh
@@ -42,10 +42,6 @@ namespace Dune
       /** \brief type of the communication */
       typedef typename Grid :: Traits :: Communication Communication;
 
-      /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-      [[deprecated("Use Communication instead!")]]
-      typedef Communication CollectiveCommunication;
-
       template< int cd >
       struct Codim
       {
@@ -97,10 +93,6 @@ namespace Dune
 
       /** \brief type of the communication */
       typedef typename Traits :: Communication Communication;
-
-      /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-      [[deprecated("Use Communication instead!")]]
-      typedef Communication CollectiveCommunication;
 
       /** \brief Codim Structure */
       template< int cd >
@@ -257,10 +249,6 @@ namespace Dune
       /** \brief type of the communication */
       typedef typename Grid :: Traits :: Communication Communication;
 
-      /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-      [[deprecated("Use Communication instead!")]]
-      typedef Communication CollectiveCommunication;
-
       template< int cd >
       struct Codim
       {
@@ -312,10 +300,6 @@ namespace Dune
 
       /** \brief type of the communication */
       typedef typename Traits :: Communication Communication;
-
-      /** \deprecated Use Communication instead! Will be removed after Dune 2.9. */
-      [[deprecated("Use Communication instead!")]]
-      typedef Communication CollectiveCommunication;
 
       /** \brief Codim Structure */
       template< int cd >