+++ CHANGELOG.md
@@ -40,6 +40,8 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 - Deprecated `CommDataHandleIF::fixedsize()` has been removed. Use `fixedSize()` with capital
   S instead.
 
+- Remove deprecated `CollectiveCommunication`, use `Communication` instead.
+
 - Deprecated `update()` member function of mappers have been removed. Use the
   member function `update(gridView)` with a grid view argument when updating
   the mapper after the grid or grid view changes. The interface